WIP: contract deployer working prototype
